+++ R/utils.R
@@ -190,7 +190,7 @@ wrap_num <- function(ggLayout, FacetWrap, FacetGrid){
 
 utils::globalVariables(c("PANEL", "axes.sequence", "density", "group",
                          "height", "positive", "setup_mapping", "x", "y",
-                         "ymax", "ymin", "fill", "..density.."))
+                         "ymax", "ymin", "fill", "..density..", "label"))
 
 get_textCoords <- function(angle, anchor, just) {